Fixed #335 (Node connection status does not update if TCP connection to gateway breaks)
Implemented an application level algorithms that plays ping pong with the other TCP endpoint assuming the TCP connection to be dead if the answer does not come within a certain time. Now, either host (gateway or portal) will be able to tell a small number of seconds if the TCP connection is still alive.

Includes one or two minor bugfixes (e.g. about the way channel pipelines were created) that became relevant with this feature/bugfix.