Resetting broken MAC address on Gateway startup and after flash operations (for #352), only tested on iSense on Mac OS X so far.
Whenever a SingleDeviceAdapter connects to a device it will check if devices' MAC address in the DeviceConfig is equal to the actual MAC address reported by the device. If not it will try to reset the device MAC address to the address configured in the DeviceConfig. The same happens when a user flashes a device. After the process the gateway will check if the device MAC is still intact and trigger the reset functionality if needed.