(Most likely) fixing ResponseTracker malfunction (#362)
Changed the RequestId logic from using random numbers to an incremental strategy. The incremental ID provider is injected as singleton and should therefore never lead to ID collisions within the portal that could lead to a malfunctioning ResponseTracker.

Also updated the gateways EventIdProvider to use the same incremental mechanism as it is done completely the same way *and* the old implementation had a bug, always generating events with the ID 0 (zero).